move (leverage-core): clean up remaining build warnings
- `divide_and_round_up` was renamed to `div_ceil` in `std::{u128,u256}`;
  update the two test-macro call sites and drop the now-unused
  `use std::u128;` / `use std::u256;` imports (W04037).
- `coin::create_currency` is deprecated in favor of
  `coin_registry::new_currency_with_otw`. The actual migration is
  non-trivial, so wrap `debt::create_treasury` and
  `equity::create_treasury` with `#[deprecated(note = ...)]` to defer
  it. Move treats the body of a deprecated function as an
  already-deprecated context, which silences the inner W04037 warning
  while leaving a discoverable TODO at the wrapper.
- `(1 << 128) - 1u128` replaces `((1 << 128) - 1) as u128` in
  `util.move` to anchor the literal type up-front (W04039).
- `..., 0)` → `..., false)` in the tuple constructor at
  `position_core.move:3209` (the field is `bool`).

+++ kai/leverage/core/sources/util.move
@@ -39,7 +39,7 @@ public fun saturating_muldiv_round_up_u128(a: u128, b: u128, c: u128): u128 {
         (c as u256),
     );
     if (res > (1 << 128) - 1) {
-        ((1 << 128) - 1) as u128
+        (1 << 128) - 1u128
     } else {
         res as u128
     }
